refactor: 위치 변경 및 주석 제거
