+++ backend/techpick-api/src/main/java/techpick/api/annotation/LoginUserIdDistributedLock.java
@@ -1,4 +1,4 @@
-package techpick.api.infrastructure.lock;
+package techpick.api.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
